requirements.txt
pysimplegui
pymongo
python-dotenv
dnspython
paho-mqtt